refactor(utils): move the ecies crypto lib into @agicash/utils
Fold ~/lib/ecies (pure ECIES over @noble) into @agicash/utils so the SDK's encryption layer
(next chunk) can use it without importing from the web app. Kept in @agicash/utils rather than
a separate package per the decision to split later only if needed.

- ecies/{ecies,index,ecies.test}.ts -> packages/utils/src/ecies (byte-identical git mv; no
  import changes — only @noble imports)
- @agicash/utils exports a "./ecies" subpath + root re-export; gains @noble/ciphers/curves/hashes
  (ciphers hoisted to the catalog; curves/hashes already there)
- old ~/lib/ecies path kept as a transitional re-export shim (its one consumer,
  shared/encryption.ts, untouched)

typecheck, biome, build all green; 135 tests (utils now 32: money + ecies).
