+++ apps/web-wallet/app/lib/ecies/index.ts
@@ -1,1 +1,2 @@
-export * from './ecies';
+// Transitional re-export — moved to @agicash/utils; removed in the import-cleanup PR.
+export * from '@agicash/utils/ecies';
