+++ apps/web-wallet/package.json
@@ -26,7 +26,7 @@
     "@agicash/utils": "workspace:*",
     "@agicash/wallet-sdk": "workspace:*",
     "@cashu/cashu-ts": "catalog:",
-    "@noble/ciphers": "1.3.0",
+    "@noble/ciphers": "catalog:",
     "@noble/curves": "catalog:",
     "@noble/hashes": "catalog:",
     "@radix-ui/react-checkbox": "1.3.3",
