+++ packages/utils/package.json
@@ -5,6 +5,7 @@
   "exports": {
     ".": "./src/index.ts",
     "./money": "./src/money/index.ts",
+    "./ecies": "./src/ecies/index.ts",
     "./*": "./src/*.ts"
   },
   "scripts": {
@@ -12,6 +13,9 @@
     "test": "bun test"
   },
   "dependencies": {
+    "@noble/ciphers": "catalog:",
+    "@noble/curves": "catalog:",
+    "@noble/hashes": "catalog:",
     "big.js": "catalog:",
     "zod": "catalog:"
   },
