+++ packages/utils/src/index.ts
@@ -2,3 +2,4 @@ export * from './money';
 export * from './json';
 export * from './zod';
 export * from './collections';
+export * from './ecies';
